+++ src/functionalTest/groovy/com/autonomousapps/AbstractFunctionalSpec.groovy
@@ -32,7 +32,7 @@ abstract class AbstractFunctionalSpec extends Specification {
   }
 
   static Boolean isV1() {
-    return System.getProperty('v') == '1'
+    return System.getProperty('dependency.analysis.old.model').toBoolean()
   }
 
   protected static void clean(ProjectDirProvider projectDirProvider) {