+++ src/functionalTest/groovy/com/autonomousapps/jvm/RippleSpec.groovy
@@ -2,7 +2,6 @@ package com.autonomousapps.jvm
 
 import com.autonomousapps.jvm.projects.RippleProject
 import org.spockframework.runtime.extension.builtin.PreconditionContext
-import spock.lang.IgnoreIf
 import spock.lang.PendingFeatureIf
 
 import static com.autonomousapps.utils.Runner.build
@@ -11,7 +10,7 @@ import static com.google.common.truth.Truth.assertThat
 final class RippleSpec extends AbstractJvmSpec {
 
   // TODO V2: Uncertain if we want to keep this feature in v2
-  @PendingFeatureIf({ PreconditionContext it -> it.sys.v == '2' })
+  @PendingFeatureIf({ PreconditionContext it -> it.sys.'dependency.analysis.old.model' == 'false' })
   def "ripples account for transitive dependencies (#gradleVersion)"() {
     given:
     def project = new RippleProject()