+++ src/functionalTest/groovy/com/autonomousapps/jvm/SimilarAdviceSpec.groovy
@@ -8,7 +8,7 @@ import static com.autonomousapps.utils.Runner.build
 import static com.google.common.truth.Truth.assertThat
 
 // TODO V2: this test can be deleted once we are fully migrated to v2
-@IgnoreIf({ PreconditionContext it -> it.sys.v == '2' })
+@IgnoreIf({ PreconditionContext it -> it.sys.'dependency.analysis.old.model' == 'false' })
 final class SimilarAdviceSpec extends AbstractJvmSpec {
 
   // https://github.com/autonomousapps/dependency-analysis-android-gradle-plugin/issues/386