+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 Dependency Analysis Plugin Changelog
 
+# Version 1.0.0-rc01 (unreleased)
+* New model is now the default. To use the old model, specify `-Ddependency.analysis.old.model=true`.
+* [Fixed] Don't report kotlin as unused when it is in fact used.
+
 # Version 0.80.0
 * [Fixed] Resolve LinkageError when more than one annotation processor references the same class.
   ([#556](https://github.com/autonomousapps/dependency-analysis-android-gradle-plugin/pull/556))